+++ packages/framework/src/Form/Constraints/ConstraintValue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Constraints;
+namespace Shopsys\FrameworkBundle\Form\Constraints;
 
 class ConstraintValue
 {
