+++ packages/framework/src/Form/Constraints/NotIdenticalToEmailLocalPartValidator.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Shopsys\FrameworkBundle\Component\Constraints;
+namespace Shopsys\FrameworkBundle\Form\Constraints;
 
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 
-class FieldsAreNotIdenticalValidator extends ConstraintValidator
+class NotIdenticalToEmailLocalPartValidator extends ConstraintValidator
 {
     /**
      * @param array $values
@@ -14,15 +14,15 @@ class FieldsAreNotIdenticalValidator extends ConstraintValidator
      */
     public function validate($values, Constraint $constraint)
     {
-        if (!$constraint instanceof FieldsAreNotIdentical) {
-            throw new \Symfony\Component\Validator\Exception\UnexpectedTypeException($constraint, FieldsAreNotIdentical::class);
+        if (!$constraint instanceof NotIdenticalToEmailLocalPart) {
+            throw new \Symfony\Component\Validator\Exception\UnexpectedTypeException($constraint, NotIdenticalToEmailLocalPart::class);
         }
 
         $propertyAccessor = PropertyAccess::createPropertyAccessor();
-        $value1 = $propertyAccessor->getValue($values, $constraint->field1);
-        $value2 = $propertyAccessor->getValue($values, $constraint->field2);
+        $password = $propertyAccessor->getValue($values, $constraint->password);
+        $email = $propertyAccessor->getValue($values, $constraint->email);
 
-        if ($value1 === $value2) {
+        if (strpos($email, $password . '@') === 0) {
             $this->context->buildViolation($constraint->message)
                 ->atPath($constraint->errorPath)
                 ->addViolation();
