+++ project-base/src/Shopsys/ShopBundle/Form/Front/Cart/CartFormType.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Form\Front\Cart;
 
-use Shopsys\FrameworkBundle\Component\Constraints\ConstraintValue;
+use Shopsys\FrameworkBundle\Form\Constraints\ConstraintValue;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
