+++ project-base/src/Shopsys/ShopBundle/Form/Front/Customer/Password/ResetPasswordFormType.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Form\Front\Customer\Password;
 
-use Shopsys\FrameworkBundle\Component\Constraints\Email;
+use Shopsys\FrameworkBundle\Form\Constraints\Email;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
