+++ project-base/src/Shopsys/ShopBundle/Form/Front/Login/LoginFormType.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Form\Front\Login;
 
-use Shopsys\FrameworkBundle\Component\Constraints\Email;
+use Shopsys\FrameworkBundle\Form\Constraints\Email;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
