+++ project-base/src/Shopsys/ShopBundle/Form/Front/Newsletter/SubscriptionFormType.php
@@ -2,7 +2,7 @@
 
 namespace Shopsys\ShopBundle\Form\Front\Newsletter;
 
-use Shopsys\FrameworkBundle\Component\Constraints\Email;
+use Shopsys\FrameworkBundle\Form\Constraints\Email;
 use Shopsys\FrameworkBundle\Form\HoneyPotType;
 use Shopsys\FrameworkBundle\Form\TimedFormTypeExtension;
 use Symfony\Component\Form\AbstractType;
