+++ project-base/src/Shopsys/ShopBundle/Form/Front/Registration/RegistrationFormType.php
@@ -2,10 +2,10 @@
 
 namespace Shopsys\ShopBundle\Form\Front\Registration;
 
-use Shopsys\FrameworkBundle\Component\Constraints\Email;
-use Shopsys\FrameworkBundle\Component\Constraints\FieldsAreNotIdentical;
-use Shopsys\FrameworkBundle\Component\Constraints\NotIdenticalToEmailLocalPart;
-use Shopsys\FrameworkBundle\Component\Constraints\UniqueEmail;
+use Shopsys\FrameworkBundle\Form\Constraints\Email;
+use Shopsys\FrameworkBundle\Form\Constraints\FieldsAreNotIdentical;
+use Shopsys\FrameworkBundle\Form\Constraints\NotIdenticalToEmailLocalPart;
+use Shopsys\FrameworkBundle\Form\Constraints\UniqueEmail;
 use Shopsys\FrameworkBundle\Form\HoneyPotType;
 use Shopsys\FrameworkBundle\Form\TimedFormTypeExtension;
 use Shopsys\FrameworkBundle\Model\Customer\UserData;
